+++ templates/base_kit_package/lib/src/ui/views/buttons/app_multiple_tap.dart
@@ -43,7 +43,7 @@ class _AppMultipleTapState extends State<AppMultipleTap> {
     });
   }
 
-  _onCompleted() {
+  void _onCompleted() {
     if (_count >= widget.taps) {
       widget.onMultipleTap.call();